Finish some missing stuff for prepared statements
All tests that have to deal with statements have been parametrized for using
both statements, the ones none prpared and the ones prepared for testing both
use cases.

Since now all signatures won't ask for bytes argyments, promoting the string
as a first way of receiving parameters and doing the encoding behind the scenes
for making the life easier to the developers.