Prepared statements support (#2)
Prepared statments are created by using the `session.create_prepared()` coroutine.
Once you have a prepared statement, servers will have them cached, a simple `statement`
can be build by just using the `prepared.bind()` function.

Other changes incorporated in this PR:

- We had to modify CallbackWrapper due to the duality of the cass_future, since
this PR the CallbackWrapper only takes the responsibility to wake up the future
and schedule the Cassandra callback. Parsing the results from Cassandra's future
becomes the responsibility of the caller.
- Added a couple of more errors. There is a lot of work to do here.
- For creating statements the `create_statement` will be the public interface, since
behind the scenes, the object is used either by raw statements and prepared statements and
the constructor can not be public.
- All tests that have to deal with statements have been parametrized for using
both statements, the ones none prepared and the ones prepared for testing both
use cases.
- Since now all signatures won't ask for bytes arguments, promoting the string
as a first way of receiving parameters and doing the encoding behind the scenes
for making life easier for the developers.