Add Python interfaces
With this PR all Cython classes are described as Python interfaces
allowing the user to know what methods each instance class has.

Also from now on any factory would need to be done through the Python
land that provides a way for closing the loop in terms of documentation
and typing, for example consider the following snippet:

    from acsylla import create_cluster, Cluster

    def my_cluster(self, hosts) -> Cluster:
        return create_cluster(hosts)

All interfaces are exposed as first citizen classes that the user can
use for typing and for getting documentation.

When an existing function/coroutine that was already there, and got
documented becuase of the addition of the interfaces, like for example
the coroutine `cluster.create_session()` has not been isolated as a
separated factory since there is an explicit method that already provides
the documentation.