Add support for fething results in a paginated way
PR provides support for tunning the page size for prepared and none
prepared statments.

Once this page size is configured, by default disabled, we can set
the page token state for fetching the next page.

Is responsability of the caller of aksing if there are more pages and
if there are set the token and repeate the operation until there is no
more pages to be fetched.

Cassandra might return a very last page with empty results when the page size is
equal to the number of results returned in the previous call.