Configure the consistency at cluster or sentence level
This PR introduces support for configuring the default level
of consistency that would be used, cluster level, and for configuring
it at statement level.

When the statement consistency is used, it will override the value provided
during the cluster configuration.

Prepared statments can be also configured for using always the same consistency level.