WIP Pythonic bind and return rows interface
Draft implementation on how the `bind`, `bind_by_name`, `bind_list` and `bind_dict` would
be exposed for replacing the current set of `bind` functions that require the type

Peformance is not impacted compared to master. As more rows to be bound less OPS/SEC, but degradation happens
in master and the proposal by the same parmeters.

TOOD
----

- Add UUID support
- Change tests
- Make the row result also more pythonic